Move customize to a sub-namespace. Change template-general to functions-template.
